:recycle: refactor: client 및 facade 패키지 구조 변경
